AddressCopier component and use it in PageHeader
